Java Exercise 2
This exercise goes over booleans that take multiple arguments.

It also goes over the "or operator" if you were to go with the answer with a lower space complexity.

There is also a much shorter answer that automatically returns true or false based off of a simplification in the logic involved.