Java Exercise 4
This is going over a function that takes a boolean and and an int as an argument.

The solution uses an if statement with a boolean, an and statement, a math operator, and a or operator as a condition to return either true or false.

There is also pseudocode at the bottom to show the process of breaking down this problem.